StateManager refactor
